RUIAN: improve DataSource bounds checking
First, we must check both new and retraced object bounds, not only the new one.

Second, avoid objects that are too close to DataSource bounds because we also
edit object's neighborhood. Minimal required area is equal to the highest
distance tolerance used in code, but it doesn't harm to ask more.