Avoid transient center and zoom to downloaded area
DownloadOsmTask automatically centers and zooms to downloaded area.
This is undesirable in case of automatic tracer downloads, so we
disable this behavior.

The trick is as follows:

(1) Create a subclass of DownloadOsmTask.DownloadTask and override
computeBboxAndCenterScale() to be a no-op.

(2) Create a subclass of DownloadOsmTask and override download(4)
method to use our own subclass for DownloadTask object.